added isExpired method to AttributeModifiers
This lets the attribute check for if a modifier is currently expired (and remove it) before using it in the final value.

+++ core/src/main/java/me/arcademadness/omni_dungeon/attributes/Armor.java
@@ -1,7 +1,10 @@
 package me.arcademadness.omni_dungeon.attributes;
 
-public final class Armor extends SimpleAttribute<Float> {
-    public Armor(float baseArmor) {
+/***
+ * Represents a flat reduction to incoming damage.
+ */
+public final class Armor extends SimpleAttribute<Integer> {
+    public Armor(int baseArmor) {
         super(baseArmor);
     }
 }
